Signed int compiler warning reduced
Squashed commit of the following:

commit 133588cc754ba130897cb53b37e156ba5c84b8fa
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Thu Aug 10 09:11:45 2023 -0400

    Updating version to 3.26.2

commit 3993d9a066fc60175e6e077d84fe8f8f0be06cec
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sun Aug 6 14:23:34 2023 -0400

    makefile

commit 555ecf2f57d8f65aac2df6c8f1c0ba4bb31ffb96
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 18:13:48 2023 +0200

    unreachable

commit 416cc86109dbdf089b4dff8422dcefbc9075e9a3
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 18:10:48 2023 +0200

    good except unused parameters, need to run unit tests

commit 4b716c3c96bbd9418a6a8b83f52d070da8a95b7c
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:54:35 2023 +0200

    mercurius almost

commit 5db1ffea2c95ae6eb9e59100abe745d9fc868c02
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:48:00 2023 +0200

    tess clean

commit 8e017fb680b9bf3c9d9ce42ea910d6482cfe381e
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:30:34 2023 +0200

    input

commit 782d9d5aa6d5914603d79055e46863fc2622c2fd
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:29:53 2023 +0200

    output good

commit ec3f36b4055d4cbdd90ca72eeaa112639f2c8f36
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:23:56 2023 +0200

    collisions free

commit 09609f2d678dbd57687b84f836922bda583a356c
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:22:11 2023 +0200

    display warning free

commit 11613c282658a2b7da7f206d4f3732ede4182574
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:21:36 2023 +0200

    tools warning free

commit c7170a7bda52369e888d4bc9b6d91ff1c4509694
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:20:42 2023 +0200

    derivatives warning free

commit f03963a3a5455ab523d6d357478dec71830cd4fb
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:17:23 2023 +0200

    simulationarchive warning free

commit 73efd634471111f5817c15b9e6020f0d77934008
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:15:53 2023 +0200

    janus warning free

commit 5641191f399d5df59b89ef6f798c12c58edbb90d
Author: Hanno Rein <hanno@hanno-rein.de>
Date:   Sat Aug 5 17:13:49 2023 +0200

    transformations without warnings

+++ src/collision.c
@@ -379,7 +379,7 @@ void reb_collision_search(struct reb_simulation* const r){
                                 c.p2--;
                             }
                         }else{
-                            if (c.p2 == r->N-r->N_var){
+                            if (c.p2 == (int)(r->N-r->N_var)){
                                 c.p2 = c.p1;
                             }
                         }
@@ -399,10 +399,10 @@ void reb_collision_search(struct reb_simulation* const r){
                                     cp->p2--;
                                 }
                             }else{
-                                if (cp->p1 == r->N-r->N_var){
+                                if (cp->p1 == (int)(r->N-r->N_var)){
                                     cp->p1 = c.p1;
                                 }
-                                if (cp->p2 == r->N-r->N_var){
+                                if (cp->p2 == (int)(r->N-r->N_var)){
                                     cp->p2 = c.p1;
                                 }
                             }
@@ -440,10 +440,10 @@ void reb_collision_search(struct reb_simulation* const r){
                                     cp->p2--;
                                 }
                             }else{
-                                if (cp->p1 == r->N-r->N_var){
+                                if (cp->p1 == (int)(r->N-r->N_var)){
                                     cp->p1 = c.p2;
                                 }
-                                if (cp->p2 == r->N-r->N_var){
+                                if (cp->p2 == (int)(r->N-r->N_var)){
                                     cp->p2 = c.p2;
                                 }
                             }
