+++ src/particle.c
@@ -130,7 +130,7 @@ void reb_add(struct reb_simulation* const r, struct reb_particle pt){
 }
 
 int reb_particle_check_testparticles(struct reb_simulation* const r){
-    if (r->N_active == r->N || r->N_active == -1){
+    if (r->N_active == (int)r->N || r->N_active == -1){
         return 0;
     }
     // Check if testparticle of type 0 has mass!=0
@@ -187,7 +187,7 @@ static struct reb_particle* reb_search_lookup_table(struct reb_simulation* const
             right = middle-1;
         }
         else if(lookuphash == hash){
-            if(lookup[middle].index < r->N){
+            if(lookup[middle].index < (int)r->N){
                 return &r->particles[lookup[middle].index];
             }
             else{ // found lookup table entry pointing beyond r->N in particles array. Needs update
@@ -208,7 +208,7 @@ static void reb_update_particle_lookup_table(struct reb_simulation* const r){
     const struct reb_particle* const particles = r->particles;
     int N_hash = 0;
     int zerohash = -1;
-    for(int i=0; i<r->N; i++){
+    for(unsigned int i=0; i<r->N; i++){
         if(N_hash >= r->allocatedN_lookup){
             r->allocatedN_lookup = r->allocatedN_lookup ? r->allocatedN_lookup * 2 : 128;
             r->particle_lookup_table = realloc(r->particle_lookup_table, sizeof(struct reb_hash_pointer_pair)*r->allocatedN_lookup);
@@ -295,9 +295,9 @@ int reb_remove(struct reb_simulation* const r, int index, int keepSorted){
     if (r->integrator == REB_INTEGRATOR_MERCURIUS){
         keepSorted = 1; // Force keepSorted for hybrid integrator
         struct reb_simulation_integrator_mercurius* rim = &(r->ri_mercurius);
-        if (rim->dcrit_allocatedN>0 && index<rim->dcrit_allocatedN){
-            for (int i=0;i<r->N-1;i++){
-                if (i>=index){
+        if (rim->dcrit_allocatedN>0 && index<(int)rim->dcrit_allocatedN){
+            for (unsigned int i=0;i<r->N-1;i++){
+                if ((int)i>=index){
                     rim->dcrit[i] = rim->dcrit[i+1];
                 }
             }
@@ -307,7 +307,7 @@ int reb_remove(struct reb_simulation* const r, int index, int keepSorted){
             struct reb_simulation_integrator_mercurius* rim = &(r->ri_mercurius);
             int after_to_be_removed_particle = 0;
             int encounter_index = -1;
-            for (int i=0;i<rim->encounterN;i++){
+            for (unsigned int i=0;i<rim->encounterN;i++){
                 if (after_to_be_removed_particle == 1){
                     rim->encounter_map[i-1] = rim->encounter_map[i] - 1; 
                 }
@@ -316,7 +316,7 @@ int reb_remove(struct reb_simulation* const r, int index, int keepSorted){
                     after_to_be_removed_particle = 1;
                 }
             }
-            if (encounter_index<rim->encounterNactive){
+            if (encounter_index<(int)rim->encounterNactive){
                 rim->encounterNactive--;
             }
             rim->encounterN--;
@@ -330,7 +330,7 @@ int reb_remove(struct reb_simulation* const r, int index, int keepSorted){
 		reb_warning(r, "Last particle removed.");
 		return 1;
 	}
-	if (index >= r->N || index < 0){
+	if (index >= (int)r->N || index < 0){
 		char warning[1024];
         sprintf(warning, "Index %d passed to particles_remove was out of range (N=%d).  Did not remove particle.", index, r->N);
 		reb_error(r, warning);
@@ -348,7 +348,7 @@ int reb_remove(struct reb_simulation* const r, int index, int keepSorted){
         if(index<r->N_active){
             r->N_active--;
         }
-		for(int j=index; j<r->N; j++){
+		for(unsigned int j=index; j<r->N; j++){
 			r->particles[j] = r->particles[j+1];
 		}
         if (r->tree_root){
